Implements new products notification task
Adds a scheduled task to send email notifications to users about newly added products.

The task runs daily at 8:30 AM and identifies products created since the last execution. It then generates an HTML email with product details and sends it to all active users in separate threads to avoid blocking.
Also adds a transactional annotation to the save method in the PedidosServiceImpl class, to ensure that the changes are atomic.
Adds a method to find all active users in the UsersRepository and UsersService classes.
Enables scheduling in the main application class.